requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
python-multipart==0.0.20
pydantic-settings==2.11.0
email-validator==2.3.0
pymongo==4.15.2
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-jose[cryptography]==3.5.0
beautifulsoup4==4.14.2
